requirements.txt
absl-py==1.4.0
astor==0.8.1
cachetools==5.3.0
certifi
charset-normalizer==3.0.1
cycler==0.11.0
fonttools==4.38.0
gast==0.2.2
google-auth==2.16.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.1
h5py==3.8.0
idna==3.4
importlib-metadata==6.0.0
joblib==0.13.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.5.3
numpy==1.21.6
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.0
pandas==1.3.5
Pillow==9.4.0
protobuf==3.20.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7.1
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
tensorboard==1.15.0
tensorboard-data-server==0.6.1
tensorboard-logger==0.1.0
tensorboard-plugin-wit==1.8.1
tensorflow==1.15.2
tensorflow-estimator==1.15.1
termcolor==2.2.0
threadpoolctl==3.1.0
# torch @ file:///media/hyone/My%20Passport/23-0210-UbuntuFile-backup/torch-1.8.1%2Bcu111-cp37-cp37m-linux_x86_64.whl
torchvision==0.9.1+cu111
tqdm==4.31.1
typing_extensions==4.4.0
urllib3==1.26.14
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.13.0
